requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-OAuthlib==0.9.6
oauthlib==3.2.2
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.8.0
SQLAlchemy==2.0.27
eventlet==0.35.1
gunicorn==21.2.0
python-engineio==4.9.0
python-socketio==5.11.1
Werkzeug==3.0.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
click==8.1.7
six==1.16.0
dnspython==2.6.1
